backend/inventario_servicio/gestor_stock/requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.0
Flask==3.1.0
flask-cors==5.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
packaging==24.2
psycopg2-binary==2.9.10
pulsar-client==3.3.0
PyJWT==2.10.1
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3